Add option to countdown in zoegraph
If the expected number of years are passed as an argument, it shows the
remaining time to reach the expected number of years. If the chronograph
surpassed the expected number of years it continues to count up.

Also this commit includes options to control precision and the refresh
interval which can be set to improve performance or higher precision.