Add preference window
Introduced a preference window for modifying the default settings of the
extension. The layout of the preference window is defined in settings.ui.
The preference window can be accessed through gnome-extension manager or
through the command line. Any changes made in the preference window become
visible after restarting the shell, and they persist across sessions.